fix logic error rejecting valid values and accepting invalid values
If the entire strtoll() string is a valid integer, endptr is NULL, not
non-null. So the reversed check meant that "3" was invalid input, but
"3.3" was valid, and so was "3.hithere" or "3junk".

Now "3.3" is considered invalid, which is fine because unlike some other
seq's out there this one doesn't do floating point (and we silently
dropped the decimal anyway soooooo...)

More importantly, "3" is actually considered valid.